server/requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
playwright>=1.30.0
playwright-stealth>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.21.0
httpx>=0.23.0
redis>=4.4.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
bcrypt>=4.0.0
cachetools>=5.0.0
loguru>=0.6.0
pytest>=7.0.0
psutil>=5.9.0